refactor(desktop): extract preset settings helpers
